move allocate_section_extents tests to test_layout_utils
